First import of our providers and services, atm twitter and identi.ca
Since we'll be using our providers and services we won't be cross
platform from day one, but this is something we'll fix with upstream
after we experiment a little bit with our own.

Upstream wants to unify this, but all pieces are not there yet.

We'll start by porting choqok so we can learn more about accounts-sso
and what we'll be the best way to make app's really cross-platform.