Make the provider & service files translatable
We'll use intltools for extracting the translatable strings for KDE's
i18n. However that needs to have the translatable tags marked with
leading underscore (eg. <_description>); these xml files are then unable
to be parsed by AccountsSSO, so we use the intltool-merge to recreate
the files with translations merged in (english), making it loadable by
AccountsSSO.

It's the same mechanism that Ubuntu uses for translations.