Rename the provider files we ship
This is a different fix for 347219 to make Ubuntu happy, but also to
make everyone else happy.

The files will now not overlap anymore, which should fix Ubuntu
packaging. The problem now got moved to the account management UIs,
which we can fix on our side if we need to.

BUG: 347219