Bundle nextcloud icon
We currently use the `Nextcloud` icon, which is installed by the official nextcloud desktop client. Therefore we don't see an icon if the client is not installed.

Instead of relying on the client being installed let's ship our own copy of the icon.

The icon is copied from https://github.com/nextcloud/desktop/blob/master/theme/colored/256-Nextcloud-icon.png. The repository is GPL-licensed so it should be fine to use.

BUG: 418745