[*cloud] Don't manually close the window when finished
We listen to the window closing to cancel the job. Since we also call
close on the window when the wizard finishes this can lead to the job
being canceled instead.

We don't need to close the window ourselves since that happens
automatically on destruction

(cherry picked from commit b9e46e8a60aeb5dbf98e2b0070b78314dd1b4351)