Port away from KPackage related KDeclarative::QmlObject methods
As discussed on the last KF6 weekly, the KPackage related methods should be phased out.

This also has the adventage that we are more flexible in case we no longer want to have
the generic KPackage structure.

A bit ugly is that we need to save the KPluginMetaData separately, but because we only have
two occurrences it should be fine.

Task: https://phabricator.kde.org/T12137